+++ packages/import-bundle/test/bundle3.js
@@ -1,0 +1,5 @@
+import { readFileSync } from 'fs';
+
+const fileContents = readFileSync('self.js', 'utf8');
+
+export default fileContents;